move helper functions from adapter.rs to request_helpers.rs
